pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "essentials-openapi"
dynamic = ["version"]
authors = [{ name = "Roberto Prevato", email = "roberto.prevato@gmail.com" }]
description = "Classes to generate OpenAPI Documentation v3 and v2, in JSON and YAML."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]
keywords = [
    "openapi",
    "docs",
    "swagger",
    "api",
    "documentation",
    "v3",
    "v2",
    "json",
    "yaml",
    "Markdown",
]

dependencies = ["PyYAML>=6", "essentials>=1.1.5", "MarkupSafe~=2.1.2"]

[tool.hatch.version]
path = "openapidocs/__init__.py"

[project.optional-dependencies]
full = ["click~=8.1.3", "Jinja2~=3.1.2", "rich~=12.6.0", "httpx<1"]

[project.scripts]
openapidocs = "openapidocs.main:main"
oad = "openapidocs.main:main"

[tool.hatch.build.targets.wheel]
packages = ["openapidocs"]

[tool.hatch.build.targets.sdist]
exclude = ["tests"]

[tool.hatch.build]
only-packages = false
include = [
    "LICENSE",
    "README.md",
    "CHANGELOG.md",
    "openapidocs/**/*.py",
    "openapidocs/**/*.html",
    "openapidocs/**/*.md",
]

[project.urls]
"Homepage" = "https://github.com/Neoteroi/essentials-openapi"
"Bug Tracker" = "https://github.com/Neoteroi/essentials-openapi/issues"
